requirements.txt
nanobot-ai==0.1.4.post5
uvicorn
django
django-cors-headers
python-dotenv
reportlab
djangorestframework-simplejwt
aiohttp
botbuilder-core
botbuilder-integration-aiohttp
python-telegram-bot
discord.py
twilio
pyngrok
slack_bolt
openpyxl
pypdf
python-docx
matplotlib
pillow
pyyaml